requirement.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
idna==2.8
mysqlclient==1.4.6
Pillow==7.0.0
pkg-resources==0.0.0
pyfcm==1.4.7
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
stripe==2.42.0
urllib3==1.25.8
